Renamed to com.labs.android to stay consistent with Firebase
