+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.awesomeproject",
+    package = "com.labs.android",
 )
 
 android_resource(
     name = "res",
-    package = "com.awesomeproject",
+    package = "com.labs.android",
     res = "src/main/res",
 )
