+++ android/app/src/main/java/com/labs/android/MainActivity.java
@@ -1,4 +1,4 @@
-package com.awesomeproject;
+package com.labs.android;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "AwesomeProject";
+        return "LabsApp";
     }
 }
