+++ android/app/src/main/java/com/labs/android/MainApplication.java
@@ -1,4 +1,4 @@
-package com.awesomeproject;
+package com.labs.android;
 
 import android.app.Application;
