+++ ios/LabsApp/Images.xcassets/AppIcon.appiconset/Contents.json
