+++ ios/LabsApp/Images.xcassets/Contents.json
