+++ ios/LabsAppTests/LabsTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface AwesomeProjectTests : XCTestCase
+@interface LabsTests : XCTestCase
 
 @end
 
-@implementation AwesomeProjectTests
+@implementation LabsTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
