+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "AwesomeProject",
+  "name": "LabsApp",
   "version": "0.0.1",
   "private": true,
   "scripts": {
